split up from_byte_reader module
